+++ Projects/Market/Sources/Screens/NotLoggedIn/LoginVideoView.swift
@@ -1,0 +1,92 @@
+import AVKit
+import Foundation
+import SwiftUI
+import UIKit
+
+struct LoginVideoView: UIViewRepresentable {
+
+    func updateUIView(_ uiView: UIView, context: UIViewRepresentableContext<PlayerView>) {
+    }
+
+    func makeUIView(context: Context) -> UIView {
+        return PlayerUIView(frame: .zero)
+    }
+}
+
+private class PlayerUIView: UIView {
+    private let playerLayer = AVPlayerLayer()
+
+    required init?(coder: NSCoder) {
+        fatalError("init(coder:) has not been implemented")
+    }
+
+    deinit {
+        NotificationCenter.default.removeObserver(self)
+    }
+
+    override init(frame: CGRect) {
+        super.init(frame: frame)
+
+        // Load the resource
+        let fileUrl = Bundle.module.url(forResource: "9x16_pillow", withExtension: "mp4")!
+
+        // Setup the player
+        let player = AVPlayer(playerItem: AVPlayerItem(url: fileUrl))
+
+        playerLayer.player = player
+        playerLayer.videoGravity = .resizeAspectFill
+        layer.addSublayer(playerLayer)
+
+        // Setup looping
+        player.actionAtItemEnd = .none
+        NotificationCenter.default.addObserver(
+            self,
+            selector: #selector(playerItemDidReachEnd(notification:)),
+            name: .AVPlayerItemDidPlayToEndTime,
+            object: player.currentItem
+        )
+
+        // Start the movie
+        player.playImmediately(atRate: 1)
+
+        NotificationCenter.default.addObserver(
+            self,
+            selector: #selector(didEnterForeground(notification:)),
+            name: UIApplication.willEnterForegroundNotification,
+            object: nil
+        )
+
+    }
+    var reversed = false
+
+    @objc
+    func didEnterForeground(notification: Notification) {
+        playerLayer.player?.play()
+    }
+    @objc
+    func playerItemDidReachEnd(notification: Notification) {
+        if let view = self.snapshotView(afterScreenUpdates: false) {
+            self.addSubview(view)
+            view.snp.makeConstraints { make in
+                make.edges.equalToSuperview()
+            }
+            UIView.animate(withDuration: 0.1, delay: 0.1) {
+                view.alpha = 0
+            } completion: { finished in
+                view.removeFromSuperview()
+            }
+            if !reversed {
+                playerLayer.player?.playImmediately(atRate: -1)
+                reversed = true
+            } else {
+                playerLayer.player?.playImmediately(atRate: 1)
+                reversed = false
+            }
+        }
+    }
+
+    override func layoutSubviews() {
+        super.layoutSubviews()
+        playerLayer.frame = bounds
+    }
+}
